Create a reusable custom hook for fetching games
